refactor tensorflow tests
