+++ .github/workflows/build.yml
@@ -134,7 +134,6 @@ jobs:
       - name: Build
         env:
           GSK_DISABLE_ANALYTICS: True
-          PYTEST_XDIST_AUTO_NUM_WORKERS: 4
         run: ./gradlew build test integrationTest :python-client:lint :python-client:package jacocoTestReport --info --parallel
 
       - name: Analyze with Sonar
